Add TemplateDeclaration::all_template_parameters
The all_template_parameters method gets the complete set of template parameters
that can affect a given item.

Note that this item doesn't need to be a template declaration itself for `Some`
to be returned from this method (in contrast to `self_template_params`). If this
item is a member of a template declaration, then the parent's template
parameters are included here. See the example in TemplateDeclaration's doc
comment for details.